requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==1.3.1
setuptools==75.7.0
tweepy==4.14.0
urllib3==2.3.0
langchain==0.0.277
schedule==1.2.0
python-dotenv==0.20.0
openai==1.59.5